refactor(db): move push store ownership
